Updates after discussion with Ken today. Removing several functions from public/utils and providing a single function to get urls.  Updated docs and tests
