--- docs/dev/code_reference/platform_mapper.md
@@ -1,5 +1,0 @@
-# Platform Mapper
-
-::: netutils.platform_mapper
-    options:
-        show_submodules: True
