+++ netutils/utils.py
@@ -93,10 +93,8 @@ _JINJA2_FUNCTION_MAPPINGS = {
     "get_upgrade_path": "os_version.get_upgrade_path",
     "hash_data": "hash.hash_data",
     "get_ips_sorted": "ip.get_ips_sorted",
-    "os_platform_object_builder": "nist.os_platform_object_builder",
-    "juniper_junos_metadata": "os_version.juniper_junos_metadata",
     "version_metadata": "os_version.version_metadata",
-    "default_metadata": "os_version.default_metadata",
+    "get_nist_urls": "nist.get_nist_urls",
 }
